+++ game-nioh3/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - None Planned
 
+## [0.2.0] - 2026-02-20
+
+- Added: Automatic install and support for DLL Loader and .dll/.asi mods
+- Added: Automatic install and support for Yumia fdata Tools and .fdata package mods - Yumia merge runs on deploy - Vanilla files restored on purge  - Credit to eArmada8 for Yumia (<https://github.com/eArmada8/yumia_fdata_tools>)
+
 ## [0.1.0] - 2026-02-04
 
 - Inital Release